Add files via upload
The assembly is comprised of three parts, the 62mmAdapterArm attaches to the sled with a screw, and the MiddleAdapter_0.6mmOffset_MFP attaches to the other end with screws on the side and bottom. Finally the MFP_2x2CoilPlate attaches to the other end of the MiddleAdapter_0.6mmOffset_MFP with screws as well (M3x10), with the the Bruker 1H receive only 2x2 mouse OR Rat surface array coil being placed on the bed of the MFP_2x2CoilPlate.